+++ src/Universalis.Application/Views/UploadCountHistoryView.cs
@@ -6,6 +6,6 @@ namespace Universalis.Application.Views
     public class UploadCountHistoryView
     {
         [JsonProperty("uploadCountByDay")]
-        public IList<uint> UploadCountByDay { get; set; }
+        public IList<double> UploadCountByDay { get; set; }
     }
 }
